split up tests
