+++ game-lobby/src/test/kotlin/TestSetup.kt
@@ -1,6 +1,11 @@
+import org.apache.kafka.clients.consumer.ConsumerRecord
+import org.apache.kafka.common.serialization.StringSerializer
 import org.apache.kafka.streams.StreamsConfig
 import org.apache.kafka.streams.TopologyTestDriver
+import org.apache.kafka.streams.test.ConsumerRecordFactory
 import java.util.*
+
+const val EMPTY_LOBBY = "{\"games\":[]}"
 
 fun setup(): TopologyTestDriver {
     // setup test driver
@@ -12,4 +17,21 @@ fun setup(): TopologyTestDriver {
     val topo = Application("dummy-brokers").build()
     println(topo.describe())
     return TopologyTestDriver(topo, props)
+}
+
+fun initLobby(testDriver: TopologyTestDriver) {
+    val factory =
+        ConsumerRecordFactory(
+            StringSerializer(),
+            StringSerializer()
+        )
+
+
+    val cr: ConsumerRecord<ByteArray, ByteArray> =
+        factory.create(
+            Topics.GAME_LOBBY_CHANGELOG,
+            GameLobby.TRIVIAL_KEY, EMPTY_LOBBY
+        )
+
+    testDriver.pipeInput(cr)
 }
